+++ binary_tree.h
@@ -6,14 +6,14 @@
 
 // Определение структуры узла бинарного дерева
 typedef struct TreeNode {
-    int value;               // Значение узла
+    float value;               // Значение узла
     struct TreeNode* left;   // Указатель на левое поддерево
     struct TreeNode* right;  // Указатель на правое поддерево
 } TreeNode;
 
 // Прототипы функций
-TreeNode* createNode(int value);           // Создание нового узла
-TreeNode* append(TreeNode* root, int value); // Добавление узла в дерево
+TreeNode* createNode(float value);           // Создание нового узла
+TreeNode* append(TreeNode* root, float value); // Добавление узла в дерево
 void preoder(TreeNode* root);               // Предзаказ обхода
 void LNR_Traversal(TreeNode* root);         // Обход в порядке LNR
 void RNL_Traversal(TreeNode* root);         // Обход в порядке RNL
